backend/requirements.txt
Django
djangorestframework
django-cors-headers
celery
redis
python-dotenv
Pillow
razorpay
reportlab